refactor: pod scaler include  hpa/vpa/manual scaler
