+++ internal/tools/orchestrator/components/podscaler/bundle.service.go
@@ -12,10 +12,14 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package components
+package podscaler
 
 import (
-	_ "github.com/erda-project/erda/internal/tools/orchestrator/components/addon/mysql"
-	_ "github.com/erda-project/erda/internal/tools/orchestrator/components/podscaler"
-	_ "github.com/erda-project/erda/internal/tools/orchestrator/components/runtime"
+	"github.com/erda-project/erda/apistructs"
 )
+
+type BundleService interface {
+	CheckPermission(req *apistructs.PermissionCheckRequest) (*apistructs.PermissionCheckResponseData, error)
+	GetCurrentUser(userID string) (*apistructs.UserInfo, error)
+	GetApp(id uint64) (*apistructs.ApplicationDTO, error)
+}
