+++ internal/tools/orchestrator/components/podscaler/common.go
@@ -1,0 +1,74 @@
+// Copyright (c) 2021 Terminus, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package podscaler
+
+import (
+	"context"
+
+	"github.com/pkg/errors"
+
+	"github.com/erda-project/erda/apistructs"
+	"github.com/erda-project/erda/internal/pkg/user"
+	"github.com/erda-project/erda/internal/tools/orchestrator/dbclient"
+	"github.com/erda-project/erda/internal/tools/orchestrator/services/apierrors"
+	"github.com/erda-project/erda/pkg/common/apis"
+	"github.com/erda-project/erda/pkg/strutil"
+)
+
+func (s *podscalerService) GetUserAndOrgID(ctx context.Context) (userID user.ID, orgID uint64, err error) {
+	orgIntID, err := apis.GetIntOrgID(ctx)
+	if err != nil {
+		err = apierrors.ErrGetRuntime.InvalidParameter(errors.New("Org-ID"))
+		return
+	}
+
+	orgID = uint64(orgIntID)
+
+	userID = user.ID(apis.GetUserID(ctx))
+	if userID.Invalid() {
+		err = apierrors.ErrGetRuntime.NotLogin()
+		return
+	}
+
+	return
+}
+
+func (s *podscalerService) checkRuntimeScopePermission(userID user.ID, runtime *dbclient.Runtime, action string) error {
+	perm, err := s.bundle.CheckPermission(&apistructs.PermissionCheckRequest{
+		UserID:   userID.String(),
+		Scope:    apistructs.AppScope,
+		ScopeID:  runtime.ApplicationID,
+		Resource: "runtime-" + strutil.ToLower(runtime.Workspace),
+		Action:   action,
+	})
+
+	if err != nil {
+		return err
+	}
+
+	if !perm.Access {
+		return apierrors.ErrGetRuntime.AccessDenied()
+	}
+
+	return nil
+}
+
+func (s *podscalerService) getUserInfo(userID string) (*apistructs.UserInfo, error) {
+	return s.bundle.GetCurrentUser(userID)
+}
+
+func (s *podscalerService) getAppInfo(id uint64) (*apistructs.ApplicationDTO, error) {
+	return s.bundle.GetApp(id)
+}
