+++ internal/tools/orchestrator/components/podscaler/db.service.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package horizontalpodscaler
+package podscaler
 
 import (
 	"github.com/erda-project/erda/internal/tools/orchestrator/dbclient"
