+++ internal/tools/orchestrator/components/podscaler/utils/apierror.go
